Update README.md
When I was working as Computer Technician, a lot of tracking of was paper based, this will include receiving requests in paper form, checking them as complete and handing them to Managment for further processing. I thought, why not put my skills to use by building a MEAN application. This by no means strictly follows the Business Process as this was done as a side project for my experience and to practice my Web Development Skills. This was used as a simple tool for my job as a Computer Technician. The objective is to track the large quantities of devices and customers I go through in a daily basis as a technician and see how much I achieved in a daily basis and overall, this will include tracking my pendings and completions of jobs and how much I completed in a monthly basis among a few extra tools.